+++ Common/Internal/Reflection/ISignatureHelpProviderWrapper.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 
@@ -6,6 +6,11 @@ namespace MirrorSharp.Internal.Reflection {
     internal interface ISignatureHelpProviderWrapper {
         bool IsTriggerCharacter(char ch);
         bool IsRetriggerCharacter(char ch);
-        Task<SignatureHelpItemsData> GetItemsAsync(Document document, int position, SignatureHelpTriggerInfoData triggerInfo, CancellationToken cancellationToken);
+        Task<SignatureHelpItemsData?> GetItemsAsync(
+            Document document, int position,
+            SignatureHelpTriggerInfoData triggerInfo,
+            SignatureHelpOptionsData options,
+            CancellationToken cancellationToken
+        );
     }
 }